pyproject.toml
[project]
name = "obfusfs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fuse-python>=1.0.9",
    "pycryptodome>=3.23.0",
]
